requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
motor>=3.7.0
pydantic-settings>=2.5.0
pydantic[email]>=2.0.0
pymongo>=4.11.0
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0